Add global endpoint rate limiter
